find_files no longer ignores untracked files
`git ls-files` only shows files that git is aware of. This causes
problems when running a freshly created test file, as it will not
be found. We could tell ls-files to include untracked files, but
that is too slow. So we add a special case to just check if the
pattern is an exact path to an existing file.