Update app root logic
We have a case where the current logic for finding the app root is not
working. It's a case with the following:

/path/to/app/subdir/test/my_test.rb

Previously, we would determine the app root to be `/path/to/app/subdir`
because we simply took the parent directory to the `test/` directory.
In the event that we had multiple test directories, then we would look
for the presence of a Gemfile|gemspec|etc.

We now always determine the app root based on the presence of the
Gemfile|gemspec|etc and fallback to the parent directory of the `test/`
directory if one is not found.

This causes a huge pain in the butt in the tests, because we now always
hit the file system. What I should do is change the "MemorySearcher" to
just actually write to the filesystem and let the tests actually search
it. But in the meantime, I've added this set of ugly hacks to write to a
tmpdir but pretend we didn't. :( Oh well, being scrappy here.